+++ blueprints/ember-pikaday/index.js
@@ -2,6 +2,6 @@ module.exports = {
   normalizeEntityName: function() {},
 
   afterInstall: function() {
-    return this.addAddonToProject('ember-cli-moment-shim', '^3.3.3');
+    return this.addAddonToProject('ember-cli-moment-shim', '^3.5.0');
   }
 };