+++ .eslintrc.js
@@ -1,10 +1,16 @@
 module.exports = {
   root: true,
+  plugins: [
+    'ember'
+  ],
   parserOptions: {
-    ecmaVersion: 6,
+    ecmaVersion: 8,
     sourceType: 'module'
   },
-  extends: 'eslint:recommended',
+  extends: [
+    'eslint:recommended',
+    'plugin:ember/recommended'
+  ],
   env: {
     browser: true
   },