+++ addon/components/pikaday-inputless.js
@@ -1,13 +1,16 @@
-import Ember from 'ember';
+import { set } from '@ember/object';
+import Component from '@ember/component';
 import PikadayMixin from 'ember-pikaday/mixins/pikaday';
 import layout from 'ember-pikaday/templates/pikaday-inputless';
 
-export default Ember.Component.extend(PikadayMixin, {
+export default Component.extend(PikadayMixin, {
   layout,
 
   didInsertElement() {
-    this.set('field', this.$('.ember-pikaday-input')[0]);
-    this.set('pikadayContainer', this.$('.ember-pikaday-container')[0]);
+    // this.set('field', this.$('.ember-pikaday-input')[0]);
+    set(this, 'field', this.$('.ember-pikaday-input')[0]);
+    // this.set('pikadayContainer', this.$('.ember-pikaday-container')[0]);
+    set(this, 'pikadayContainer', this.$('.ember-pikaday-container')[0]);
     this.setupPikaday();
   },